feat(ui): map tap-as-hover, persistent overlay, remove CtRegionMapDebug (#1045)
- Map widget: tap on mobile drives hover selector and province glow (SPEC/ui/map-widget.md).
- Province overlay: stays open until close button; no toggle on re-tap (SPEC/ui/province-sea-zone-detail-overlay.md).
- Remove CtRegionMapDebug; CtRegionMap is the single implementation.
- Move CtMapVisibilityMode to region_map_component; update imports, widgetbook, tests, widget_catalog and specs.

Made-with: Cursor
