+++ app/test/province_overlay_test.dart
@@ -7,7 +7,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'package:colonizethis_app/features/game/widgets/province_sea_zone_detail_overlay.dart';
 import 'package:colonizethis_app/features/game/widgets/province_overlay_demo_data.dart';
-import 'package:colonizethis_app/widgets/region_map_debug.dart';
+import 'package:colonizethis_app/widgets/ct_region_map.dart';
 
 void main() {
   suppressLogsForTests();
@@ -313,7 +313,7 @@ void main() {
             body: Row(
               children: [
                 Expanded(
-                  child: CtRegionMapDebug(
+                  child: CtRegionMap(
                     region: demoRegionForOverlay,
                     cellSizePx: 28,
                     onProvinceSelected: (_) {},
@@ -336,13 +336,13 @@ void main() {
           ),
         ),
       );
-      await tester.pumpAndSettle();
+      await tester.pump();
 
-      expect(find.byType(CtRegionMapDebug), findsOneWidget);
+      expect(find.byType(CtRegionMap), findsOneWidget);
       expect(find.byType(ProvinceSeaZoneDetailOverlay), findsOneWidget);
     });
 
-    testWidgets('AC: Map tap invokes onProvinceSelected; overlay can show selection',
+    testWidgets('AC: Map tap invokes onProvinceSelected; overlay can show selection and stays open until closed',
         (WidgetTester tester) async {
       final region = demoRegionForOverlay;
       String? selectedId;
@@ -356,7 +356,7 @@ void main() {
                     SizedBox(
                       width: 400,
                       height: 320,
-                      child: CtRegionMapDebug(
+                      child: CtRegionMap(
                         region: region,
                         cellSizePx: 28,
                         onProvinceSelected: (id) =>
@@ -383,23 +383,26 @@ void main() {
           ),
         ),
       );
-      await tester.pumpAndSettle();
+      await tester.pump();
 
       expect(selectedId, isNull);
-      final mapFinder = find.byType(CtRegionMapDebug);
-      final element = tester.element(
-        find.descendant(
-          of: mapFinder,
-          matching: find.byType(SizedBox),
-        ).first,
-      );
-      final box = element.renderObject! as RenderBox;
-      final center = box.localToGlobal(box.size.center(Offset.zero));
-      await tester.tapAt(center);
-      await tester.pumpAndSettle();
+      final mapFinder = find.byType(CtRegionMap);
+      await tester.tap(mapFinder);
+      await tester.pump();
 
       expect(selectedId, isNotNull);
       expect(selectedId!, startsWith('${region.regionId}|'));
+
+      // Tapping the same province again should not hide the overlay; it remains
+      // visible until the close handler clears the selection. SPEC/ui/province-sea-zone-detail-overlay.md.
+      await tester.tap(mapFinder);
+      await tester.pump();
+      expect(selectedId, isNotNull);
+
+      // Close via the overlay and verify it clears selection.
+      await tester.tap(find.byIcon(Icons.close));
+      await tester.pump();
+      expect(selectedId, isNull);
     });
   });
 }
